compiler warning fixes
--HG--
branch : HEAD

+++ src/lib-storage/index/maildir/maildir-sync-index.c
@@ -452,7 +452,7 @@ int maildir_sync_index(struct maildir_index_sync_context *ctx,
 
 	if (uid_validity == 0) {
 		uid_validity = hdr->uid_validity != 0 ?
-			hdr->uid_validity : ioloop_time;
+			hdr->uid_validity : (uint32_t)ioloop_time;
 		maildir_uidlist_set_uid_validity(mbox->uidlist, uid_validity);
 	}
 	maildir_uidlist_set_next_uid(mbox->uidlist, hdr_next_uid, FALSE);
